Realized I had all my code in MainActivity so I added monsterActivity
